flight_replay/requirements.txt
streamlit
pandas
plotly
networkx
pyserial
PyQt6
pyqtgraph
